Refactor geography helper to geography
